server/requirements.txt
Flask==0.10.1
PyPubSub==3.3.0
bitalino==0.6
numpy==1.6.2
pyserial==2.7
